Move _clear_rft_caches fixture to ert/conftest.py
Commit 6b6ad2bc8eadafc29231bd4eb3e26f5d1e3f7d6c made the
mocked_files fixture available to all ert tests but left the
_read_egrid/_get_zonemap cache-clearing behind in test_rft_config.py.
